requirements.txt
streamlit>=1.28.0
opencv-python-headless==4.8.1.78
numpy>=1.24.0
pillow>=10.0.0
plotly>=5.15.0
scipy>=1.11.0